Industrial-training-IIUC/Industrial training IIUCSession1_part2/DB implement with scrapper/requirements.txt
mysql-connector-python
python-dotenv
requests-html
pandas
numpy
